pyproject.toml
[tool.poetry]
name = "shop-hopper"
version = "1.1.2"
description = "Script for finding desired products across multiple marketplaces to compare availability and prices"
authors = ["Pavel Soroka <40a.pvl@gmail.com>"]
readme = "README.md"


[tool.poetry.scripts]
shop-hopper = "shop_hopper.scripts.app:main"


[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.2"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
python-dotenv = "^1.0.1"
jinja2 = "^3.1.4"
selenium = "^4.26.1"
webdriver-manager = "^4.0.2"
rich = "^14.0.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
